+++ ErsatzTV.Core/Interfaces/Metadata/Nfo/IMovieNfoReader.cs
@@ -1,0 +1,8 @@
+﻿using ErsatzTV.Core.Metadata.Nfo;
+
+namespace ErsatzTV.Core.Interfaces.Metadata.Nfo;
+
+public interface IMovieNfoReader
+{
+    Task<Either<BaseError, MovieNfo>> Read(Stream input);
+}