+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Fix adding episodes with no title to the search index
+  - This behavior was preventing some items from being removed from the trash
 
 ## [0.5.3-beta] - 2022-04-29
 ### Fixed