+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - This behavior was preventing some items from being removed from the trash
 - Support combination NFO metadata for movies
   - Note that ErsatzTV does not scrape any metadata; any URLs after the XML will be ignored
+- Fix bug causing some Jellyfin and Emby content to incorrectly show as unavailable
 
 ### Added
 - Use `Sort Title` from Movie NFO metadata if available